fix: use custom hook for infinity scroll
